Product detail view quantity select (#24)
* Add to cart function

Added the ability to add to cart from the Product Detail View. Also gives an message when an item is added to the cart. Currently this message displays regardless of success or failure.

* Moved Snackbar to ProductCard

Moved the snack bar message to the Product Card class so that it shows only when necessary.